+++ services/backend/toolwarehouse-service/src/main/java/com/veteroch4k/toolwarehouse/repositories/FreeToolsRepository.java
@@ -3,6 +3,6 @@ package com.veteroch4k.toolwarehouse.repositories;
 import com.veteroch4k.toolwarehouse.models.FreeTools;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FreeToolsRepository extends JpaRepository<FreeTools, Long> {
+public interface FreeToolsRepository extends JpaRepository<FreeTools, Integer> {
 
 }
