+++ services/backend/toolwarehouse-service/src/main/java/com/veteroch4k/toolwarehouse/repositories/ToolAccountingRepository.java
@@ -3,6 +3,6 @@ package com.veteroch4k.toolwarehouse.repositories;
 import com.veteroch4k.toolwarehouse.models.ToolAccounting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToolAccountingRepository extends JpaRepository<ToolAccounting, Long> {
+public interface ToolAccountingRepository extends JpaRepository<ToolAccounting, Integer> {
 
 }
