+++ services/backend/toolwarehouse-service/src/main/java/com/veteroch4k/toolwarehouse/repositories/ToolRepository.java
@@ -5,7 +5,7 @@ import com.veteroch4k.toolwarehouse.models.ToolType;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToolRepository extends JpaRepository<Tool, Long> {
+public interface ToolRepository extends JpaRepository<Tool, Integer> {
 
   List<Tool> findToolsByToolType_Name(String type);
